- sorting stats in query rather than code - moving datastore reports to report namespace rather than member
